split scenes and sceneTypes out of sceneLogic
